+++ common/src/main/java/tocraft/walkers/mixin/PlayerManagerMixin.java
@@ -38,7 +38,7 @@ public class PlayerManagerMixin {
                 player.getAttribute(Attributes.ARMOR)
                         .setBaseValue(Math.min(Walkers.CONFIG.maxAmor, shape.getAttributeBaseValue(Attributes.ARMOR)));
                 player.getAttribute(Attributes.ARMOR_TOUGHNESS)
-                        .setBaseValue(Math.min(Walkers.CONFIG.maxAmor, shape.getAttributeBaseValue(Attributes.ARMOR_TOUGHNESS)));
+                        .setBaseValue(Math.min(Walkers.CONFIG.maxAmorToughness, shape.getAttributeBaseValue(Attributes.ARMOR_TOUGHNESS)));
             }
             // sync max health & attack damage with clients
             if (Walkers.CONFIG.scalingHealth || Walkers.CONFIG.percentScalingHealth || Walkers.CONFIG.scalingAmor) {